Added more code cleaning and enhacements
