+++ packages/camera/src/components/UploadCenter/hooks/index.js
@@ -1,0 +1,3 @@
+export { default as useComplianceIds } from './useComplianceIds';
+export { default as useHandlers } from './useHandlers';
+export { default as useMixedStates } from './useMixedStates';
